강학생 신청한것만 보이는 화면

+++ src/main/java/com/green/greenuni/application/lectures/model/MyLectureListReq.java
@@ -9,7 +9,7 @@ import lombok.ToString;
 @ToString
 
 public class MyLectureListReq {
-    private int lectureId;
+    private Long lectureId;
     private Long loginUserId;
     private int page;
     private int size;
